7.5 fix multiple touches navigation
